% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/combine_contacts.R
\name{combine_contacts}
\alias{combine_contacts}
\title{Combine two contact matrices in one, split by the diagonal}
\usage{
combine_contacts(x1, x2, common.scale = F, diag.value = NA)
}
\arguments{
\item{x1}{First contact matrix (upper diagonal)}

\item{x2}{Second contact matrix (lower diagonal)}

\item{common.scale}{Logical indicating if both matrices should be on the same scale}

\item{diag.value}{Scaler or vecor of values to susbtitute the diagonal elements by}
}
\value{
A squared matrix of contacts
}
\description{
This function takes two contact matrices and combines them into a new pseudo-contact matrix with one experiment in the upper diagonal and the other in the lower diagonal
}
\examples{
plot(0)
}
\seealso{
\code{\link{read_tabix}} and \code{\link{get_contacts}} for data retrieval
}

