% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get_contacts.R
\name{get_contacts}
\alias{get_contacts}
\title{Create a contact matrix from a data frame containing a list of contacts.}
\usage{
get_contacts(matrix_file, region)
}
\arguments{
\item{matrix_file}{Input data frame. Usually the output of \code{\link{read_tabix}}}

\item{region}{Region of interest (chr:start-end)}
}
\value{
numeric matrix with the number of contacts per pair of loci and genomic coordinates as dimnames
}
\description{
Create a contact matrix from a data frame containing a list of contacts.
}
\seealso{
\code{\link{read_tabix}} for data retrieval
}

