% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/make_contacts.R
\name{make_contacts}
\alias{make_contacts}
\title{Get contact matrix from compressed and indexed reads}
\usage{
make_contacts(infile, region, window)
}
\arguments{
\item{infile}{Input tab separated, compressed and indexed contact's read file (chr1 pos1 strand1 chr2 pos2 strand2 read_ID)}

\item{region}{Region to retrieve the contacts from. Accepts expression wit the format "chr:start-position", "chr:start" or "chr"}

\item{window}{Resolution in bp}
}
\value{
A squared matrix of contacts
}
\description{
This function takes a compressed and indexed file of contact reads and returns a contact matrix
}
\examples{
plot(0)
}
\seealso{
\code{\link{read_tabix}} and \code{\link{get_contacts}} for data retrieval
}

