% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contacts.R
\name{plot_contacts}
\alias{plot_contacts}
\title{Plot a HiC map from a contact matrix.}
\usage{
plot_contacts(contacts, rotate = T, transformation = log10,
  col.palette = brewer.pal(9, "Blues"))
}
\arguments{
\item{contacts}{Input contact matrix. Usually the output of \code{\link{get_contacts}}}

\item{rotate}{Do you want matrix to be rotated so the diagonal is from top-left to bottom-right? Dafaults to TRUE}

\item{transformation}{Transformation of the data to ease the visualization. Dafaults to log10}

\item{col.palette}{Palette of colors used in the representation}
}
\value{
A nice plot
}
\description{
This function takes a contact matrix and plots a HiC map
}
\examples{
plot(0)
}
\seealso{
\code{\link{read_tabix}} and \code{\link{get_contacts}} for data retrieval
}

