% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tabix.R
\name{read_tabix}
\alias{read_tabix}
\title{Slice and read a region from a tabix-inedxed file.}
\usage{
read_tabix(thefile, theregion)
}
\arguments{
\item{thefile}{Input file name}

\item{theregion}{Region of interest (chr:start-end)}
}
\value{
data frame with the information of the region of interest.
}
\description{
Slice and read a region from a tabix-inedxed file.
}
\seealso{
\code{\link{tabix.read}} which this function wraps
}

