% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCompare.R, R/runSimple.R
\name{runSimple}
\alias{runSimple}
\title{Launch a shiny app to visualize HiC data.}
\usage{
runSimple()

runSimple()
}
\description{
Launch a shiny app to visualize HiC data.

Launch a shiny app to visualize HiC data.
}

