% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_contacts_rotated.R
\name{plot_contacts_rotated}
\alias{plot_contacts_rotated}
\title{Plot a rotated HiC map from a contact matrix.}
\usage{
plot_contacts_rotated(mat, m1 = NULL, m2 = NULL,
  transformation = function(x) log10(x + 0.9))
}
\arguments{
\item{mat}{Input contact matrix. Usually the output of \code{\link{get_contacts}} or \code{\link{make_contacts}}}

\item{m1}{Label of the lower panel}

\item{m2}{Label of the upper panel}

\item{transformation}{Transformation of the data to ease the visualization. Dafaults to log10}
}
\value{
A nice plot
}
\description{
This function takes a contact matrix and plots a HiC map
}
\examples{
plot(0)
}
\seealso{
\code{\link{read_tabix}} and \code{\link{get_contacts}} for data retrieval
}

